Ensure full mapping of paths to container
The step in `CommandLineTool.job` that implicitly maps files in
builder.bindings to container paths via dictionary assignment
misses some cases in complex inputs. The result is that builder.bindings
will have correct internal Docker paths in `valueFrom` but not in `do_eval`,
leading to command line arguments that still reference the external
path.

To fix, this does an explicit `adjustFiles` remapping over all elements
in builder.bindings, supplementing the work done by the implicit call
and ensuring both `builder.files` and `builder.bindings` are
synchronized.

Thanks to @tetron for discussion and pointers on digging into the code
to fix this.