Enable src/sink comparisons for records
The logic for comparing records from two different sections of the
workflow previous failed due to falling back to a `==` comparison that
would not pass due to differences in fully qualified record names.

This fix adds an explicit record comparison function that normalizes
field names by the record namespace and then ensures that all of the
types of each field are identical from both records.

Also adds a test to demonstrate the behavior.