Move invocation of output callback for workflows from the job() method to the receive_output() method.
This closes a gap whereby sometimes all steps of a subworkflow have completed,
but it has not yet called its output callback.  This can lead to the parent
workflow incorrectly believing no progress has been made and returning
'None' (which indicates a workflow deadlock if there are no jobs currently
executing) instead of re-evaluating pending steps to see if they can be run.