Fix #411: Incorrect length of input with few sources
Fix #411
If the source for step input has more than one type,
it's added to result array for this input exactly
the same number of times as the lenght of type list
for this source.
In function 'workflow.py/match_types' we don't need
to check for all source types to be compatible with
the sink, 'cos the source could be, for example, int
and null, but the sink is only int. But it shouldn't
cause an error, 'cos I may use the case when the
source is null in valueFrom field. The reason why
it doesn't fail right now is the bug in here
https://github.com/common-workflow-language/cwltool/
blob/972b0a6d7c2e8abfb287ddedeee48938e76689c9/cwltool
/workflow.py#L73
We should use sinktype instead of st.
The solution is to check if at least one source type
is compatible with the sink and than in linkMerge
section additionally check for can_assign_src_to_sink