Fix #687
bind_input does two things: marks files as inputs to a tool and composes what the command should. When using bind_input with a valueFrom, this code (rightly) ignores composing of the command to delegate this to another function. This also doesn't find files, which results in the error in #687. This commit fixes this.