Executor clears any saved result when execute is called. (#1813)
The executor interface is instanced rather than a constructor interface, and the most recent results are saved on the object as final_output and final_status.  We should be able to invoke the executor more than once, however the current executor will not overwrite previously recorded results with new results.  This change clears final_output and final_status each time execute() is called so that new results can be recorded.