Fixed race condition which happens when a job runs "too fast",
When a podman process finishes even before reaching the monitoring
method, a deadlock happens, as no one is updating `process.returncode`
and spawned process is in zombie state (so, no signal is sent).

This fix adds a `process.poll()` call, so it gives the chance to fill in
`process.returncode`.